+++ ipaas/src/layouts/AppLayout.tsx
@@ -106,7 +106,9 @@ import {
   X,
 } from '@wso2/oxygen-ui-icons-react';
 import FeaturePreviewModal from '../components/FeaturePreview/FeaturePreviewModal';
-import { useProject, useProjectByHandler, useProjects, useComponents, useOrgs } from '../api/queries';
+import { useProject, useProjectByHandler, useProjects } from '../hooks/useProjects';
+import { useComponents } from '../hooks/useComponents';
+import { useOrgs } from '../hooks/useOrg';
 import { SUPPORTED_DISPLAY_TYPES, GENERIC_SERVICE_TYPES } from '../constants/integrations';
 import { fetchOrgPermissions } from '../api/auth';
 import { switchOrgToken } from '../auth/tokenManager';
@@ -422,10 +424,6 @@ function AppLayoutInner(): JSX.Element {
         return 'build';
       case 'deploy':
         return hasComponent(targetScope) ? 'deploy' : 'overview';
-      case 'metrics':
-        return 'metrics';
-      case 'runtimes':
-        return 'runtimes';
       case 'environments':
         return 'environments';
     }
