+++ ipaas/src/nav.ts
@@ -33,7 +33,7 @@ export type Scope = OrgScope | ProjectScope | ComponentScope;
 
 export type ScopeForLevel = { organizations: OrgScope; projects: ProjectScope; components: ComponentScope };
 
-export type Resource = 'overview' | 'logs' | 'alerts' | 'metrics' | 'runtimes' | 'environments' | 'access-control' | 'build' | 'deploy';
+export type Resource = 'overview' | 'logs' | 'alerts' | 'environments' | 'access-control' | 'build' | 'deploy';
 
 export type Matrix = { [R in Resource]: { segment: string; pages: Partial<{ [L in Level]: FC<ScopeForLevel[L]> }> } };
