+++ ipaas/src/pages/AccessControl.tsx
@@ -22,7 +22,8 @@ import { useParams, useNavigate } from 'react-router';
 import { useAccessControl } from '../contexts/AccessControlContext';
 import { ALL_USER_MGT_PERMISSIONS, Permissions } from '../constants/permissions';
 import { componentAccessControlUrl } from '../paths';
-import { useProjectByHandler, useComponentByHandler } from '../api/queries';
+import { useProjectByHandler } from '../hooks/useProjects';
+import { useComponentByHandler } from '../hooks/useComponents';
 import type { ComponentScope } from '../nav';
 import { Loading } from './access-control/shared';
 import { UsersTab } from './access-control/UsersTab';
